src/requirements.txt
Flask==2.0.2
requests==2.26.0
xmltodict==0.12.0
Pillow==8.4.0
beautifulsoup4==4.10.0
Werkzeug==2.2.2
